Integrated Oliver Charles's CSS selector support
Changed things around, though, with just one additional argument that can be a string or a code refernce. This opens things up to the possibility of other filters going forward. Note that the recursive testing is not quite right, though, due to a but in HTML::Selector::XPath (RT #50131). Will likely delay release till I hear from Tatsuhiko-San.